+++ addons/scene_manager/sub_section.gd
@@ -1,5 +1,5 @@
 @tool
-extends Node
+extends Control
 
 # Nodes
 @onready var button: Button = find_child("Button")
@@ -11,9 +11,12 @@ const _close = preload("res://addons/scene_manager/icons/GuiOptionArrowRight.png
 # If it is "All" subsection, open it
 func _ready() -> void:
 	button.text = name
+	if name == "All" && get_child_count() == 0:
+		visible = false
 
 # Add child
 func add_item(item: Node) -> void:
+	item._sub_section = self
 	list.add_child(item)
 
 # Open list
@@ -36,3 +39,17 @@ func _on_button_up():
 		close()
 	else:
 		open()
+
+# When a node adds
+func child_entered():
+	if name == "All" && get_child_count() == 0:
+		visible = false
+	else:
+		visible = true
+
+# When a node removes
+func child_exiting():
+	if name == "All" && get_child_count() - 2 == 0:
+		visible = false
+	else:
+		visible = true